feat: Auth Logic
- Add authentication

Issue #4
